Clarify implementation & API differences
