+++ example.rb
@@ -1,24 +1,39 @@
 class ProductTask < MaintenanceTasks::Task
-  include MaintenanceTasks::Adapters::ActiveRecord
-
+  # This could be a child class instead of module, but would need an
+  # "abstract_class" mechanism like ActiveRecord itself.
+  include MaintenanceTasks::ActiveRecordTask
+  # Consumers implement these method
   def collection() Product.all end
   def process(product) product.discontinue! end
+
+  # .count implementation provided, but could be overriden if consumer has some
+  # optimization or if the relation is too expensive to count
+
+  # Could easily add
+  #   run_in_batches_of 50
+  # or
+  #   def batch_size() 50 end
 end
 
 class SeedTask < MaintenanceTasks::Task
-  include MaintenanceTasks::Adapters::Array
+  include MaintenanceTasks::ArrayTask
 
+  # Similar API to ActiveRecordTask
   def collection() %w[Alice Bob Chantalle] end
   def process(name) Person.create!(name: name) end
+  # Again, .count provided
 end
 
 class BetaFlagTask < MaintenanceTasks::Task
-  include MaintenanceTasks::Adapters::CSV
+  include MaintenanceTasks::CSVTask
 
+  # CSV infra handles everything except .process
   def process(row) BetaFlag[:whatever].enable(row.shop_id) end
 end
 
 class PaymentTask < MaintenanceTasks::Task
+  # Consumers implement .enumerator instead of .collection, without having to
+  # know that other tasks actually also use .enumerator internally
   def enumerator(cursor:)
     Enumerator.new do |yielder|
       loop do
@@ -32,6 +47,7 @@ class PaymentTask < MaintenanceTasks::Task
     end
   end
 
+  # Again, can be implemented if appropriate
   def count() InvoiceAPI.unpaid_count end
 
   def process(invoice) invoice.pay! if invoice.unpaid? end
