--- lib/maintenance_tasks/adapters/csv.rb
@@ -1,3 +1,0 @@
-module MaintenanceTasks::Adapters::CSV
-  # ...
-end
